Revert "Merge branch 'master' of https://github.com/GeorgianaElena/Monitoring_clarisse"
This reverts commit 7d2077a4caadf62b014b35f5e04990847d2ae873, reversing
changes made to 5960811d1fb22468af820fae61c9ab91c481ccdc.

+++ metrics_aggregator.h
@@ -70,7 +70,7 @@ typedef struct _metrics_t
   long timestamp;
   int quit;
 #ifdef BENCHMARKING
-  double start_time;
+  long start_time;
 #endif
 } metrics_t, *metrics_t_ptr;
 
@@ -99,7 +99,7 @@ typedef struct _metrics_aggregator_t
   node_state_t current_state;
 
 #ifdef BENCHMARKING
-  double *benchmarking_results;
+  long *benchmarking_results;
 #else
   sys_metric_t *monitoring_results;
 #endif
